test: cover keybinding widget lifecycle
