--- packages/e2e/src/keybindings.edit.ts
@@ -1,20 +1,0 @@
-import type { Test } from '@lvce-editor/test-with-playwright'
-
-export const name = 'keybindings.edit'
-
-export const test: Test = async ({ expect, KeyBindingsEditor, Locator }) => {
-  // arrange
-  await KeyBindingsEditor.open()
-  const keyBindingsView = Locator('.Viewlet.KeyBindings')
-  await expect(keyBindingsView).toBeVisible()
-  const input = Locator('.KeyBindingsSearchInputBox')
-  await expect(input).toBeVisible()
-  await KeyBindingsEditor.handleInput('About.focus')
-
-  // act
-  await KeyBindingsEditor.handleDoubleClick(70, 70)
-
-  // assert
-  const defineKeyBindingWidget = Locator('.Viewlet.DefineKeyBinding')
-  await expect(defineKeyBindingWidget).toBeVisible()
-}
