+++ packages/server/package.json
@@ -11,6 +11,6 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "@lvce-editor/server": "0.94.7"
+    "@lvce-editor/server": "0.95.7"
   }
 }
